refactor: extract verification logic into blsag/verify.rs
Move verify_with_progress and VerifyRef trait impl into dedicated
verify module. Completes the sign/verify separation in the blsag
module hierarchy.
